Backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
langchain
langchain-core
pydantic
